Return promise from `setCurrentCurrency` thunk
The `setCurrentCurrency` thunk performed async operations, but it
didn't return a Promise. It now returns a Promise.

The callers in this case never needed to know when this action had
completed, but at least this makes our tests more reliable. They were
already `await`-ing this action, despite the lack of Promise.