Return Promise from `signTx` thunk
The `signTx` action creator would return a thunk that performed async
work, but it didn't return a Promise. It now does.

This does result in a few slight changes in behavior:
* The next confirmation page is no longer shown immediately; it's shown
after the operation completes instead.
* The "Transactions / Edit Screen / Complete" metric event is now
dispatched _after_ the operation completes, and only upon a successful
opertaion. I believe this was the original intended behaviour.
Previously it would dispatch this event before the operation had
completed, even in cases where it had failed.