Return Promise from `addToAddressBook`
`addToAddressBook` had returned a thunk that performed async
operations, but it didn't return a Promise. Now it does.

The callers of this action creator were updated to `await` the
completion of the operation. It was called just before redirecting the
user to a different page or closing a modal, and it seemed appropriate
to wait before doing those things.