Return Promise from `approvePermissionsRequest` thunk
`approvePermissionsRequest` returned a thunk that didn't return a
Promise, despite doing async work. It now returns a Promise.

This action creator was only used in two places; one where it looks
like it was expected to be non-blocking, and another where it was the
last step of an event handler. So no updates were required to the
callers.