Return Promise from `clearPermissions` thunk
`clearPermissions` returned a thunk that didn't return a Promise,
despite doing async work. It now returns a Promise.

The one caller was updated to `await` the result of this operation, as
as there was a UI state update following the call that seemed like it
expected the operation to have been completed.